types

+++ ConcreteSupply.SqlSchema/dbo/Tables/ConcretePumpJobCard.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[ConcretePumpJobCard] (
-    [Id]                     INT NOT NULL,
-    [ConcretePumpHoursStart] INT NOT NULL,
-    [ConcretePumpHoursEnd]   INT NOT NULL,
+    [Id]                     INT            NOT NULL,
+    [ConcretePumpHoursStart] DECIMAL (8, 2) NOT NULL,
+    [ConcretePumpHoursEnd]   DECIMAL (8, 2) NOT NULL,
     CONSTRAINT [PK_ConcretePumpJobCard] PRIMARY KEY CLUSTERED ([Id] ASC),
     CONSTRAINT [FK_ConcretePumpJobCard_VehicleJobCard] FOREIGN KEY ([Id]) REFERENCES [dbo].[VehicleJobCard] ([Id])
 );
